pyproject.toml
[project]
name = "tg-bot-avito"
version = "0.1.0"
description = "Add your description here"
readme = "Readme.adoc"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.23.0",
    "aiohttp>=3.13.2",
    "asyncpg>=0.31.0",
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "nest-asyncio>=1.6.0",
    "psycopg2-binary>=2.9.11",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "yookassa>=3.8.0",
]
